PENDING: iris: fix to handle drain substates during streamon capture
To resume decoder/encoder operation client can call either
a pair of streamoff and streamon on the ports or issue a start
command. It is necessary to clear substates related to drain/drain_last
in both sequences. In the current code, these sub states were not
cleared in the streamon sequence. The fix clears drain related
states on streamon of the capture port and also resumes the
input port if it is in streaming state.

Change-Id: I7b9940f2fc91bf3e03ebe3b355ed0352e27c9450
Signed-off-by: Chandrakant I Viraktamath <quic_civirakt@quicinc.com>