QCLINUX: soc: qcom: Make SoC debug module as builtin module
The SoC debug module keeps the SoC watchdog device node enabled
when Linux is booted in hyp mode. To do this, it needs to update
status property of the watchdog device node. This must happen before
platform devices are populated in of_platform_default_populate_init()
which runs at arch_initcall level. Currently the SoC debug module
runs at pure_initcall level, so it is able to keep the device node
disabled before platform devices are populated. Since this is a hard
requirement, make SoC debug module as builtin module so that it can't
be compiled as a loadable module.

Change-Id: I53e3db00a31afb67b0f6db32645fe91edd13ead5
Signed-off-by: Pavankumar Kondeti <quic_pkondeti@quicinc.com>