QCLINUX: usb: dwc3: qcom: Avoid PM Resume if device is RT suspended
During XO exit use case, it has been observed that upon connection,
the role switch notification to device mode is not guaranteed to
come before PM Resume kicking in as both these are independent and
thus can come in any order.

As per the test case, before DUT enters XO, the usb device is
autp suspended. When cable is connected and glink smem wakes up
the system, there are two things that happen (not in order):

1. Role switch notification to device mode
2. PM resume kicking in and invoking drivers pm_resume operation

In the event role switch kicks in, there are two scenarios seen:
a) PPM sends role-none followed by role-device
b) PPM sends role-device directly

In both these scenarios, there is a runtime_get_sync being done
and runtime_resume kicks in bringing up successful enumeration.

But in the event PM resume kicks in, the flattened device tree
implementation invokes dwc3_qcom_pm_resume which will invoke
dwc_resume. As part of this sequence, the gadget_resume invokes
run_stop and we set the RS bit to '1' and its corresponding
vendor hook sets the qscratch OTG_VBUS_VALID to '1'. Although the
RS bit and OTG_VBUS_VALID bit are set, enumeration is not seen. At
this point, since the autosuspend is still kept enabled, usb
re-enters runtime suspend. And disabling auto suspend or doing a
Plug-In/Plug-Out is helping fix enumeration.

Since it is observed that runtime resume path is the only path that
is bringing successful enumeration (same as in downstream), bail out
in PM resume call if the device was runtime suspended and the device
tree has been flattened essentially enforcing that a RT resume should
be done to counter a RT suspend and PM Resume to be done to counter a
PM suspend.

PM suspend would only happen if done forcibly because as long as the
enumeration is present in device mode, usb won't allow DUT to enter
PM suspend. In host mode, since the cable is connected, wakeup from
PM suspend by triggering autoresume from connected device would result
in runtime resume and hence would be a valid case.

If PM suspend is done forcibly in device mode with cable connected, or
if the PM suspend entry is triggered before auto suspend kicks in (which
is not allowed in downstream), then upon PM resume, although RS is set,
device mode enumeration will fail. Banking on that fact that this is a
less used case, the PM resume is ignored if RT suspend has happened.

Change-Id: I1fb8d068ca9a3ff9549601b2196d110041a9bbb9
Signed-off-by: Krishna Kurapati <quic_kriskura@quicinc.com>