FROMLIST: PCI/pwrctl: add PCI power control core code
Some PCI devices must be powered-on before they can be detected on the
bus. Introduce a simple framework reusing the existing PCI OF
infrastructure.

The way this works is: a DT node representing a PCI device connected to
the port can be matched against its power control platform driver. If
the match succeeds, the driver is responsible for powering-up the device
and calling pcie_pwrctl_device_set_ready() which will trigger a PCI bus
rescan as well as subscribe to PCI bus notifications.

When the device is detected and created, we'll make it consume the same
DT node that the platform device did. When the device is bound, we'll
create a device link between it and the parent power control device.

Change-Id: Ie829322173d01de31e7de854d8f16416dcab01b1
Signed-off-by: Bartosz Golaszewski <bartosz.golaszewski@linaro.org>
Link: https://lore.kernel.org/linux-pci/20240410124628.171783-13-brgl@bgdev.pl/
Patch-mainline: linux-pci @ 04/10/2024, 14:46
Signed-off-by: Miaoqing Pan <quic_miaoqing@quicinc.com>