usb: gadget: Do not handle OS Descriptors config separately
Consider a multi-configuration composition in which the last
configuration uses OS descriptors. Since this configuration will
be sent first, the host may choose this config if it matches the
choosing criteria and ignore the user configured order of the
multi-config composition. Moreover linux based hosts do not re-order
the received configurations based on their indices but process them
in FIFO order. This may result in the host never choosing the user
desired configuration because of the re-order caused by sending the
OS descriptor config first. To avoid this, follow the user configured
order and do not handle OS descriptor config separately.

Change-Id: I60a65a0bf1d57ec479a05612833f7e1974a77334
Signed-off-by: Elson Roy Serrao <quic_eserrao@quicinc.com>