QCLINUX: arm64: dts: qcom: Remove iommu enc dec data from venus node
Video venus driver design is based on encoder and decoder sub nodes,
while the video iris vpu driver is designed to work as a single node.
To enable iris vpu driver as an add on, existing subnodes are deleted.
Additionally, the video iris driver supports multiple context bank
approach for IOMMU in contrast to single IOMMU in case of venus driver.
Hence the change also removes IOMMU property so that context bank
details can be added in the overlay DTSI file.

Change-Id: I7617a2046c2118de54d30be6cf9d4213d1dd8627
Signed-off-by: Chandrakant I Viraktamath <quic_civirakt@quicinc.com>