QCLINUX: usb: dwc3: core: Set driver data before core init
In latest builds, it is observed that during probe, if the
dr_mode is host, the xhci layer tries to execute plat_start
quirk which inturn accesses driver data of core to do required
register operations. In case the driver data is not set before
core_init, the driver data read by supplied plat_start quirk
will be NULL and dereferencing it will cause a crash.

Set driver data of non-flattened implementations before core
init to avoid this crash.

Change-Id: Ia5f86d97611b19ee86900ad721d9d44e657ebe98
Signed-off-by: Krishna Kurapati <quic_kriskura@quicinc.com>