BACKPORT: UPSTREAM: net: stmmac: dwmac-qcom-ethqos: Configure host DMA width
Commit 070246e4674b ("net: stmmac: Fix for mismatched host/device DMA
address width") added support in the stmmac driver for platform drivers
to indicate the host DMA width, but left it up to authors of the
specific platforms to indicate if their width differed from the addr64
register read from the MAC itself.

Qualcomm's EMAC4 integration supports only up to 36 bit width (as
opposed to the addr64 register indicating 40 bit width). Let's indicate
that in the platform driver to avoid a scenario where the driver will
allocate descriptors of size that is supported by the CPU which in our
case is 36 bit, but as the addr64 register is still capable of 40 bits
the device will use two descriptors as one address.

Change-Id: Id61572c1c62c65eebdd998e1a3200805f7534f95
Fixes: 8c4d92e82d50 ("net: stmmac: dwmac-qcom-ethqos: add support for emac4 on sa8775p platforms")
Reviewed-by: Simon Horman <horms@kernel.org>
Link: https://lore.kernel.org/all/jtalwaityx7fyakigggyahhhor23fml76yic3e3xkeoimdqoj2@i7fiqzacowq3/
Reviewed-by: Andrew Halaney <ahalaney@redhat.com>
Git-commit: 0579f27249047006a818e463ee66a6c314d04cea
Git-repo: git://git.kernel.org/pub/scm/linux/kernel/git/torvalds/linux.git
Signed-off-by: Sagar Cheluvegowda <quic_scheluve@quicinc.com>