PENDING: usb: dwc3: Redesign flattening implementation
With the recent addition of xhci plat quirks in dwc3_host_init,
there came up a new problem. If dr_mode is host, then in that case,
xhci_plat_start quirk requires dwc reg access to modify GUSB2PHYCFG
and GUSB3PIPECTL. There are two issues with this:

1. If the device tree is not flattened, then dwc platform data is
set after dwc3 probe call is done. So during host_init accessing dwc
driver data and dereferencing leads to a crash.

2. If the device tree is flattened, then the driver data points to
dwc3_qcom which is not accessible in host driver file.

To avoid this, make the callers of dwc3_probe to allocate the dwc
structure themselves. By doing so, for unflattened implementations,
the platform data can be set well in advance so that xhci_plat_start
can access it.

For flattened implementation, set the driver data to statically present
dwc3 struct within allocated dwc3_qcom so that qcom data can be obtained
by a container of call and dwc can be obtained by checking driver data.

Change-Id: Iec34dd1f2829f1291c3962a28dee2d8432ffbb8d
Signed-off-by: Krishna Kurapati <quic_kriskura@quicinc.com>