FROMLIST: misc: fastrpc: Fix fastrpc_map_lookup operation
Fastrpc driver creates maps for user allocated fd buffers. Before
creating a new map, the map list is checked for any already existing
maps using map fd. Checking with just map fd is not sufficient as the
user can pass offsetted buffer with less size when the map is created
and then a larger size the next time which could result in memory
issues. Check for user passed VA and length also when looking up for
the map.

Change-Id: I77b43f457ee1e49580f76bef079b9f9a9c793826
Link: https://patchwork.kernel.org/project/linux-arm-msm/patch/20240822105933.2644945-3-quic_ekangupt@quicinc.com/
Signed-off-by: Ekansh Gupta <quic_ekangupt@quicinc.com>