QCLINUX: qcom: tz_log: differentiate hlos and hypversior of creating shmbridge
The tz_log driver fail to create shmbridge buffer for hlos that
cause fail to create tz_log device node. This is because the
qtee_shmbridge driver register shmbridge buffer for hypervisor
and hlos with the attribute "qcom, support-hypervisor", so tz_log
must implement the same logic to differentiate hypervisor and hlos.

Change-Id: Ie454ae809d966ee683a39714eb36093a6cdf3806
Signed-off-by: Jian Shu <quic_jianshu@quicinc.com>