QCLINUX: firmware: qcom_scm_addon: Fix qcom_she_op missing arginfo and res
There are 4 args in the desc struct and arg1, arg2, arg4 are
QCOM_SCM_VAL type, arg3 is QCOM_SCM_RW type, so correct desc.arginfo
to pass the right value and buffer to the TZ side. Additionally,
the result from SHE service must be handled by the SHE client, so fix
the missing res too.

Change-Id: Ibc01f4f82531b3ffbc26ed9c20ed9939c82e7508
Signed-off-by: Jian Shu <quic_jianshu@quicinc.com>